Fixed a breaking bug
In `common.hpp` the `load_be` and `load_le` functions were broken. Instead
of reading in bytes in their respective order, these functions would output
the last byte it was supposed to read with the rest set to 0xff's. I fixed
this by explicitly casting the value read from the buffer to uint8_t, since
that's what the buffer (should) contain. This fixed the issue. I am not
quite sure why this happened though..

+++ src/common.hpp
@@ -50,7 +50,7 @@ T load_be(const char* buffer, unsigned int bufsize, unsigned int offset, unsigne
 
 	for (unsigned int i = 0; i < num; i++) {
 		temp <<= 8;
-		temp |= buffer[offset + i];
+		temp |= (u8)(buffer[offset + i]);
 	}
 	temp <<= (sizeof(T) - num) * 8;
 	return temp;
@@ -76,7 +76,7 @@ T load_le(const char* buffer, unsigned int bufsize, unsigned int offset, unsigne
 	unsigned int j = num - 1;
 	for (unsigned int i = 0; i < num; i++) {
 		temp <<= 8;
-		temp |= buffer[offset + (j--)];
+		temp |= (u8)(buffer[offset + (j--)]);
 	}
 	return temp;
 }
